server/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.2
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-oauth2-provider==0.2.6.1
django-oauth-toolkit==2.2.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.2.0
pycparser==2.21
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
oauth2_provider==0.0
django-crontab==0.7.1
requests==2.28.0
google-api-python-client==2.74.0
pyjwt==2.6.0
drf-yasg==1.21.5